requirements.txt
alembic==0.9.5
astroid==1.5.3
backports.functools-lru-cache==1.4
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
coverage==4.0.3
enum34==1.1.6
Flask==0.12.2
Flask-API==1.0b2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
psycopg2==2.6.2
pycparser==2.18
PyJWT==1.5.2
pylint==1.7.2
python-coveralls==2.9.1
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
PyYAML==3.12
requests==2.18.4
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.2.0b2
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
